spatial_point_pattern: use proj4string<- function rather than the slot *
The use of 'proj4string' in R code is confusing in times where PROJ.4 strings are obsoleted.
It can be expected that a more general function name to set an object's CRS will sooner or later
be promoted instead of proj4string<-, hence preparing for that already by using the function instead of the slot.
I expect the proj4string slot will most probably remain in use for compatibility reasons, but
explicitly naming it is better avoided in times where WKT2 strings are actually used to define a CRS.