Revert "spatial_mgrs: fix csl filepath" and drop the csl file *
This reverts commit a5f67dd11fc02ddbb31f39a585529da5ad899a5e and drops the CSL file again
(added in b78e67a).

I had tricked myself into believing that the YAML of index.md is being used (and it
contains the temp path to the csl file). However it isn't and index.md is already the
result of applying the CSL style, so no CSL file lookup needed here.